requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
openai>=1.30.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
plotly>=5.20.0
statsmodels>=0.14.0
